backend/requirements.txt
# Flask==3.1.0
python-dotenv==1.0.1
# flask-cors==5.0.1
pymongo==4.11.2
bcrypt==4.3.0
jwt==1.3.1
# flask-restful==0.3.10
# flask-pymongo==3.0.1

fastapi
python-dotenv
httpx
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
h11==0.12.0
importlib-metadata==4.6.4
pydantic==1.8.2
pydantic[email]
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
zipp==3.5.0
email-validator
python-multipart==0.0.20
aiofiles